requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytz
httpx
dateparser

# Google API Auth and Client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# OpenAI & LangChain
openai
langchain
langgraph

# Optional: If you plan to use Gemini via Google
google-generativeai
